requirements.txt
Django>=2.0,<3.0
django-phonenumber-field
phonenumbers
django-crispy-forms
django-environ
pdfkit
django-imagekit